+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.1.1] - 2020-09-25
+### Added
+- Added ability item type and Abilities tab to character sheet.
+### Fixed
+- Fixed labels on classic character sheet.
+
 ## [1.1.0] - 2020-09-23
 ### Added
 - Added Quickstart 2E character sheet support (default).